pyproject.toml
[project]
name = "six-sigma"
version = "0.1.0"
description = "Simple web app to evaluate a process with the \"6 Sigma\" approach, https://six-sigma.containerapps.ru/"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "logfire[fastapi]>=3.19.0",
    "matplotlib>=3.10.1",
    "mplcyberpunk>=0.7.6",
    "pytest-cov>=6.0.0",
    "python-decouple>=3.8",
    "scipy>=1.15.2",
]
